+++ src/main/java/com/epam/hospital/dao/StaffPatientDao.java
@@ -7,7 +7,7 @@ public interface StaffPatientDao extends Dao<StaffPatient> {
 
     int patientsOfStaffCount(long staffId) throws DBException;
 
-    int staffOfPatientCount(int patientId) throws DBException;
+    int staffOfPatientCount(long patientId) throws DBException;
 
-    void delete(int patientId) throws DBException;
+    void delete(long patientId) throws DBException;
 }
